chore: Further refactor to avoid `startedPayments` calls
