+++ .vscode/launch.json
@@ -40,7 +40,18 @@
     {
       "type": "node",
       "request": "launch",
-      "name": "Test HOPR core Ethereum indexer",
+      "name": "Test HOPR core",
+      "skipFiles": ["<node_internals>/**"],
+      "program": "${workspaceFolder}/node_modules/.bin/mocha",
+      "console": "integratedTerminal",
+      "args": ["--reporter=tap", "--full-trace", "--exit"],
+      "cwd": "${workspaceFolder}/packages/core",
+      "outFiles": ["${workspaceFolder}/**/*.js"]
+    },
+    {
+      "type": "node",
+      "request": "launch",
+      "name": "Test HOPR core-ethereum",
       "skipFiles": ["<node_internals>/**"],
       "program": "${workspaceFolder}/node_modules/.bin/mocha",
       "args": ["--reporter=tap", "--full-trace", "--exit"],
