+++ packages/core-ethereum/src/ethereum.mock.ts
@@ -1,0 +1,7 @@
+export const sampleChainOptions = {
+  chainId: 1337,
+  environment: 'hardhat-localhost',
+  gasPrice: 1.1,
+  network: 'hardhat',
+  provider: 'http://localhost:8545'
+}
