+++ packages/core-ethereum/src/ethereum.ts
@@ -25,6 +25,22 @@ const abiCoder = new utils.AbiCoder()
 export type Receipt = string
 export type ChainWrapper = Awaited<ReturnType<typeof createChainWrapper>>
 
+export class ChainWrapperSingleton {
+  private static instance: ChainWrapper;
+  private constructor() { }
+  public static async create(
+    networkInfo: { provider: string; chainId: number; gasPrice?: number; network: string; environment: string },
+    privateKey: Uint8Array,
+    checkDuplicate: Boolean = true
+  ): Promise<ChainWrapper> {
+    log('Receiving create request for hopr-ethereum with `networkInfo`', networkInfo)
+    if (!ChainWrapperSingleton.instance) {
+      ChainWrapperSingleton.instance = await createChainWrapper(networkInfo, privateKey, checkDuplicate);
+    }
+    return ChainWrapperSingleton.instance;
+  }
+}
+
 export async function createChainWrapper(
   networkInfo: { provider: string; chainId: number; gasPrice?: number; network: string; environment: string },
   privateKey: Uint8Array,
