+++ packages/core-ethereum/src/index.spec.ts
@@ -1,5 +1,6 @@
 import { dbMock, debug, privKeyToPeerId, PublicKey } from '@hoprnet/hopr-utils'
 import HoprEthereum from '.'
+import { sampleChainOptions } from './ethereum.mock'
 
 const namespace = 'hopr:test:hopr-ethereum'
 const log = debug(namespace)
@@ -10,7 +11,7 @@ const mockPeerId = privKeyToPeerId(privateKey)
 describe(`HoprEthereum`, function () {
   it('should instantiate a new class w/o any issues', function () {
     log('starting new instance of HoprEthereum.')
-    new HoprEthereum(dbMock, PublicKey.fromPeerId(mockPeerId), mockPeerId.privKey.marshal())
+    new HoprEthereum(dbMock, PublicKey.fromPeerId(mockPeerId), mockPeerId.privKey.marshal(), sampleChainOptions)
     log('successfully created the HoprEthereum instance.')
   })
 })
