+++ packages/core-ethereum/src/indexer/index.spec.ts
@@ -1,192 +1,13 @@
-import { providers as Providers, Wallet } from 'ethers'
-import type { HoprChannels, HoprToken } from '@hoprnet/hopr-ethereum'
-import type { Event, TokenEvent } from './types'
-import type { ChainWrapper } from '../ethereum'
+import { BigNumber } from 'ethers'
 import assert from 'assert'
-import EventEmitter from 'events'
-import Indexer from '.'
-import { Address, ChannelEntry, Hash, HoprDB, generateChannelId, ChannelStatus, defer } from '@hoprnet/hopr-utils'
+import { ChannelEntry, Hash, ChannelStatus, defer } from '@hoprnet/hopr-utils'
+
 import { expectAccountsToBeEqual, expectChannelsToBeEqual } from './fixtures'
 import * as fixtures from './fixtures'
 import { PARTY_A, PARTY_B } from '../fixtures'
-import { BigNumber } from 'ethers'
+import type { Event } from './types'
+import { useFixtures } from './index.mock'
 
-const txRequest = {
-  to: fixtures.ACCOUNT_B.address,
-  data: '0x0',
-  value: 0,
-  nonce: 0,
-  gasPrice: 1
-}
-
-const createProviderMock = (ops: { latestBlockNumber?: number } = {}) => {
-  let latestBlockNumber = ops.latestBlockNumber ?? 0
-
-  const provider = new EventEmitter() as unknown as Providers.WebSocketProvider
-  provider.getBlockNumber = async (): Promise<number> => latestBlockNumber
-
-  return {
-    provider,
-    newBlock() {
-      latestBlockNumber++
-      provider.emit('block', latestBlockNumber)
-    }
-  }
-}
-
-const createHoprChannelsMock = (ops: { pastEvents?: Event<any>[] } = {}) => {
-  const pastEvents = ops.pastEvents ?? []
-  const channels: any = {}
-  const pubkeys: any = {}
-
-  const handleEvent = (ev) => {
-    if (ev.event == 'ChannelUpdated') {
-      const updateEvent = ev as Event<'ChannelUpdated'>
-
-      const eventChannelId = generateChannelId(
-        Address.fromString(updateEvent.args.source),
-        Address.fromString(updateEvent.args.destination)
-      )
-      channels[eventChannelId.toHex()] = updateEvent.args.newState
-    } else if (ev.event == 'Announce') {
-      pubkeys[ev.args.account] = ev.args.multiaddr
-    } else {
-      //throw new Error("MISSING EV HANDLER IN TEST")
-    }
-  }
-
-  class FakeChannels extends EventEmitter {
-    async channels(channelId: string) {
-      for (let ev of pastEvents) {
-        handleEvent(ev)
-      }
-      return channels[channelId]
-    }
-
-    async bumpChannel(_counterparty: string, _comm: string) {
-      let newEvent = {
-        event: 'ChannelUpdated',
-        transactionHash: '',
-        blockNumber: 3,
-        transactionIndex: 0,
-        logIndex: 0,
-        args: {
-          source: PARTY_B.toAddress().toHex(),
-          destination: PARTY_A.toAddress().toHex(),
-          newState: {
-            balance: BigNumber.from('3'),
-            commitment: Hash.create(new TextEncoder().encode('commA')).toHex(),
-            ticketEpoch: BigNumber.from('1'),
-            ticketIndex: BigNumber.from('0'),
-            status: 2,
-            channelEpoch: BigNumber.from('0'),
-            closureTime: BigNumber.from('0')
-          }
-        } as any
-      } as Event<'ChannelUpdated'>
-      handleEvent(newEvent)
-      this.emit('*', newEvent)
-    }
-
-    async queryFilter() {
-      return pastEvents
-    }
-  }
-
-  const hoprChannels = new FakeChannels() as unknown as HoprChannels
-
-  return {
-    hoprChannels,
-    pubkeys,
-    newEvent(event: Event<any>) {
-      pastEvents.push(event)
-      hoprChannels.emit('*', event)
-    }
-  }
-}
-const createHoprTokenMock = () => {
-  class FakeToken extends EventEmitter {
-    async transfer() {
-      let newEvent = {
-        event: 'Transfer',
-        transactionHash: '',
-        blockNumber: 8,
-        transactionIndex: 0,
-        logIndex: 0,
-        args: {
-          source: PARTY_A.toAddress().toHex(),
-          destination: PARTY_B.toAddress().toHex(),
-          balance: BigNumber.from('1')
-        } as any
-      } as TokenEvent<'Transfer'>
-      this.emit('*', newEvent)
-    }
-  }
-
-  const hoprToken = new FakeToken() as unknown as HoprToken
-
-  return {
-    hoprToken,
-    newEvent(event: Event<any>) {
-      hoprToken.emit('*', event)
-    }
-  }
-}
-
-const createChainMock = (
-  provider: Providers.WebSocketProvider,
-  hoprChannels: HoprChannels,
-  hoprToken: HoprToken,
-  account?: Wallet
-): ChainWrapper => {
-  return {
-    getLatestBlockNumber: () => provider.getBlockNumber(),
-    subscribeBlock: (cb) => provider.on('block', cb),
-    subscribeError: (cb) => {
-      provider.on('error', cb)
-      hoprChannels.on('error', cb)
-    },
-    subscribeChannelEvents: (cb) => hoprChannels.on('*', cb),
-    subscribeTokenEvents: (cb) => hoprToken.on('*', cb),
-    getNativeTokenTransactionInBlock: (_blockNumber: number, _isOutgoing: boolean = true) => [],
-    updateConfirmedTransaction: (_hash: string) => {},
-    unsubscribe: () => {
-      provider.removeAllListeners()
-      hoprChannels.removeAllListeners()
-    },
-    getChannels: () => hoprChannels,
-    getWallet: () => account ?? fixtures.ACCOUNT_A,
-    setCommitment: (counterparty: Address, commitment: Hash) =>
-      hoprChannels.bumpChannel(counterparty.toHex(), commitment.toHex()),
-    getAllQueuingTransactionRequests: () => [txRequest]
-  } as unknown as ChainWrapper
-}
-
-const useFixtures = async (ops: { latestBlockNumber?: number; pastEvents?: Event<any>[] } = {}) => {
-  const latestBlockNumber = ops.latestBlockNumber ?? 0
-  const pastEvents = ops.pastEvents ?? []
-
-  const db = HoprDB.createMock()
-  const { provider, newBlock } = createProviderMock({ latestBlockNumber })
-  const { hoprChannels, newEvent } = createHoprChannelsMock({ pastEvents })
-  const { hoprToken } = createHoprTokenMock()
-  const chain = createChainMock(provider, hoprChannels, hoprToken)
-  return {
-    db,
-    provider,
-    newBlock,
-    hoprChannels,
-    newEvent,
-    indexer: new Indexer(Address.fromString(fixtures.ACCOUNT_A.address), db, 1, 5),
-    chain,
-    OPENED_CHANNEL: await ChannelEntry.fromSCEvent(fixtures.OPENED_EVENT, (a: Address) =>
-      Promise.resolve(a.eq(PARTY_A.toAddress()) ? PARTY_A : PARTY_B)
-    ),
-    COMMITTED_CHANNEL: await ChannelEntry.fromSCEvent(fixtures.COMMITTED_EVENT, (a: Address) =>
-      Promise.resolve(a.eq(PARTY_A.toAddress()) ? PARTY_A : PARTY_B)
-    )
-  }
-}
 
 describe('test indexer', function () {
   it('should start indexer', async function () {
