+++ packages/cover-traffic-daemon/src/index.ts
@@ -4,14 +4,14 @@ import BN from 'bn.js'
 import yargs from 'yargs/yargs'
 import { terminalWidth } from 'yargs'
 
-import { createHoprNode, resolveEnvironment, supportedEnvironments } from '@hoprnet/hopr-core'
-import { ChannelEntry, privKeyToPeerId, PublicKey, debug } from '@hoprnet/hopr-utils'
+import { createHoprNode } from '@hoprnet/hopr-core'
+import { ChannelEntry, privKeyToPeerId, PublicKey, debug, resolveEnvironment, supportedEnvironments, ResolvedEnvironment } from '@hoprnet/hopr-utils'
 
 import { PersistedState } from './state'
 import { CoverTrafficStrategy } from './strategy'
 
 import type PeerId from 'peer-id'
-import type { HoprOptions, ResolvedEnvironment } from '@hoprnet/hopr-core'
+import type { HoprOptions } from '@hoprnet/hopr-core'
 import type { PeerData, State } from './state'
 import http from 'http'
 import restana from 'restana'
