+++ packages/utils/src/index.ts
@@ -14,3 +14,4 @@ export * from './db'
 export * from './db.mock'
 export * from './ethereum'
 export * from './utils'
+export * from './environment'
